+++ examples/typescript/src/tatum.ts
@@ -1,3 +1,3 @@
-import { TatumSdk } from '@tatumcom/js'
+import { TatumSDK } from '@tatumcom/js'
 
-export const initTatum = async () => await TatumSdk.init()
+export const initTatum = async () => await TatumSDK.init({ rpc: { ignoreLoadBalancing: true }, verbose: true })