+++ src/service/index.ts
@@ -1,2 +1,3 @@
 export * from './notification'
 export * from './tatum'
+export * from './rpc'