+++ src/service/rpc/rpc.ts
@@ -1,0 +1,213 @@
+import { Container, Inject, Service } from 'typedi'
+import { CONFIG, Constant, Utils } from '../../util'
+import { TatumConnector } from '../../connector/tatum.connector'
+import { Blockchain, JsonRpcCall, JsonRpcResponse } from '../../dto'
+
+interface RpcInvocation {
+  callRpc(request: JsonRpcCall): Promise<JsonRpcResponse>
+}
+
+interface RpcStatus {
+  node: {
+    url: string
+  }
+  lastBlock: number
+  lastResponseTime: number
+  failed: boolean
+}
+
+@Service({
+  factory: (data: { id: string }) => {
+    return new Rpc(data.id)
+  }, transient: true,
+})
+export class Rpc {
+  private readonly rpcUrlMap = new Map<Blockchain, RpcStatus[]>()
+  private readonly activeUrl = new Map<Blockchain, { url: string, index: number }>()
+  private interval: NodeJS.Timeout
+
+  public readonly bitcoin: RpcInvocation
+  public readonly litecoin: RpcInvocation
+  public readonly ethereum: RpcInvocation
+  public readonly polygon: RpcInvocation
+  public readonly monero: RpcInvocation
+  @Inject()
+  readonly connector: TatumConnector
+
+  constructor(private readonly id: string) {
+    this.bitcoin = this.create(Blockchain.BITCOIN)
+    this.litecoin = this.create(Blockchain.LITECOIN)
+    this.polygon = this.create(Blockchain.POLYGON)
+    this.ethereum = this.create(Blockchain.ETHEREUM)
+    this.monero = this.create(Blockchain.MONERO)
+  }
+
+  async init() {
+    const config = Container.of(this.id).get(CONFIG)
+    if (config.rpc?.ignoreLoadBalancing) {
+      if (config.verbose) {
+        console.debug(new Date().toISOString(), 'Initializing RPC module from static URLs')
+      }
+      // we are ignoring load balancing, getting static URLs from the chains they provided configuration for
+      for (const blockchain of Object.values(Blockchain)) {
+        const urls = config.rpc[blockchain]?.url
+        if (urls?.length) {
+          this.activeUrl.set(blockchain, { url: urls[0], index: -1 })
+          if (config.verbose) {
+            console.debug(new Date().toISOString(), `Using static URL ${urls[0]} for ${blockchain}`)
+          }
+        }
+      }
+    } else {
+      if (config.verbose) {
+        console.debug(new Date().toISOString(), 'Initializing RPC module from remote hosts')
+      }
+      if (config.rpc?.oneTimeLoadBalancing && config.rpc?.waitForFastestNode) {
+        await this.checkStatus()
+      } else if (!config.rpc?.oneTimeLoadBalancing) {
+        this.interval = setInterval(() => this.checkStatus(), Constant.OPEN_RPC.LB_INTERVAL)
+        process.on('exit', () => clearInterval(this.interval))
+      }
+    }
+  }
+
+  private async checkStatus() {
+    const { verbose } = Container.of(this.id).get(CONFIG)
+    const allChains = []
+    for (const blockchain of Object.values(Blockchain)) {
+      if (!this.rpcUrlMap.has(blockchain)) {
+        continue
+      }
+      const servers: RpcStatus[] = this.rpcUrlMap.get(blockchain) as RpcStatus[]
+      const all = []
+      for (const server of servers) {
+        all.push(Utils.fetchWithTimeout(server.node.url, {
+          method: 'POST',
+          body: JSON.stringify(Constant.OPEN_RPC.STATUS_PAYLOAD[blockchain]),
+        }).then(async ({ response: res, responseTime }) => {
+          server.lastResponseTime = responseTime
+          const response = await res.json()
+          if (verbose) {
+            console.debug(new Date().toISOString(), `Response time of ${server.node.url} for ${blockchain} blockchain is ${server.lastResponseTime}ms with response: `, response)
+          }
+          if (res.ok && response.result) {
+            server.failed = false
+            server.lastBlock = Utils.statusPayloadExtractor(blockchain, response)
+          } else {
+            if (verbose) {
+              console.warn(new Date().toISOString(), `Failed to check status of ${server.node.url} for ${blockchain} blockchain.`, response)
+            }
+            server.failed = true
+          }
+        }).catch((e) => {
+          if (verbose) {
+            console.warn(new Date().toISOString(), `Failed to check status of ${server.node.url} for ${blockchain} blockchain.`, e)
+            console.warn(new Date().toISOString(), `Server ${server.node.url} for ${blockchain} will be marked as failed and will be removed from the pool.`)
+          }
+          server.failed = true
+        }))
+      }
+      allChains.push(Promise.allSettled(all).then(() => {
+        const { fastestServer, index } = Rpc.getFastestServer(servers)
+        if (fastestServer && index !== -1) {
+          if (verbose) {
+            console.debug(new Date().toISOString(), `Server ${fastestServer.node.url} for ${blockchain} blockchain is the active server.`)
+          }
+          this.activeUrl.set(blockchain, { url: fastestServer.node.url, index })
+        }
+      }))
+    }
+    await Promise.allSettled(allChains)
+  }
+
+  private create(blockchain: Blockchain): RpcInvocation {
+    const { verbose } = Container.of(this.id).get(CONFIG)
+    return {
+      callRpc: (request: JsonRpcCall) => callRpc(request, this),
+    }
+
+    async function callRpc(request: JsonRpcCall, rpc: Rpc): Promise<JsonRpcResponse> {
+      if (!rpc.rpcUrlMap.has(blockchain)) {
+        await rpc.initChainMap(blockchain)
+      }
+      const url = rpc.getActiveUrl(blockchain)
+      try {
+        return await rpc.connector.rpcCall(url, request)
+      } catch (e) {
+        const activeIndex = rpc.getActiveIndex(blockchain)
+        if (activeIndex !== -1) {
+          /// we are using static URL, so we don't need to do anything
+          throw  e
+        }
+        if (verbose) {
+          console.warn(new Date().toISOString(), `Failed to call RPC ${request.method} on ${url} for ${blockchain} blockchain.`, e)
+          console.log(new Date().toISOString(), `Switching to another server for ${blockchain} blockchain, marking this as unstable.`)
+        }
+        const servers = rpc.rpcUrlMap.get(blockchain) as RpcStatus[]
+        servers[activeIndex].failed = true
+        const { index, fastestServer } = Rpc.getFastestServer(servers)
+        if (index === -1) {
+          console.error(`All servers for ${blockchain} blockchain are unavailable.`)
+          throw e
+        }
+        rpc.activeUrl.set(blockchain, { url: fastestServer.node.url, index })
+        return callRpc(request, rpc)
+      }
+    }
+  }
+
+  private static getFastestServer(servers: RpcStatus[]) {
+    const { fastestServer, index } = servers.reduce((result, item, index) => {
+      if (!item.failed &&
+        (item.lastBlock > result.fastestServer.lastBlock ||
+          (item.lastBlock === result.fastestServer.lastBlock && item.lastResponseTime < result.fastestServer.lastResponseTime))) {
+        return { fastestServer: item, index: index };
+      } else {
+        return result
+      }
+    }, { fastestServer: { lastBlock: -Infinity, lastResponseTime: Infinity, node: { url: '' } }, index: -1 })
+    return { fastestServer, index }
+  }
+
+  private getActiveUrl(blockchain: Blockchain): string {
+    if (this.activeUrl.has(blockchain)) {
+      return this.activeUrl.get(blockchain)?.url as string
+    }
+    throw new Error(`No active URL for ${blockchain} blockchain`)
+  }
+
+  private getActiveIndex(blockchain: Blockchain): number {
+    if (this.activeUrl.has(blockchain)) {
+      return this.activeUrl.get(blockchain)?.index as number
+    }
+    throw new Error(`No active URL for ${blockchain} blockchain`)
+  }
+
+  private async initChainMap(blockchain: Blockchain) {
+    const { verbose } = Container.of(this.id).get(CONFIG)
+    if (verbose) {
+      console.debug(new Date().toISOString(), `Fetching response from ${Constant.OPEN_RPC.CONFIG_URL[blockchain]}`)
+    }
+    try {
+      const res = await fetch(Constant.OPEN_RPC.CONFIG_URL[blockchain])
+      if (res.ok) {
+        const servers: Array<{ url: string }> = await res.json()
+        const randomIndex = Math.floor(Math.random() * servers.length)
+        if (verbose) {
+          console.debug(new Date().toISOString(), `Using random URL ${servers[randomIndex].url} for ${blockchain} blockchain during the initialization`)
+        }
+        this.activeUrl.set(blockchain, { url: servers[randomIndex].url, index: randomIndex })
+        this.rpcUrlMap.set(blockchain, servers.map((s) => ({
+          node: { url: s.url },
+          lastBlock: 0,
+          lastResponseTime: 0,
+          failed: false,
+        })))
+      } else {
+        console.error(new Date().toISOString(), `Failed to fetch RPC configuration for ${blockchain} blockchain`)
+      }
+    } catch (e) {
+      console.error(new Date().toISOString(), 'Failed to initialize RPC module', e)
+    }
+  }
+}