+++ src/service/tatum/tatum.dto.ts
@@ -1,9 +1,84 @@
 export interface TatumConfig {
-    network?: Network
-    validate?: boolean
-    debug?: boolean
-    retryDelay?: number,
-    retryCount?: number
+  /**
+   * Blockchain to use, mainnet or testnet. Mainnet is the default one.
+   */
+  network?: Network
+
+  /**
+   * Verbose logging is disabled by default.
+   */
+  verbose?: boolean
+
+  /**
+   * Delay between retries, defaults to 1000ms.
+   */
+  retryDelay?: number,
+
+  /**
+   * Number of retries in case of failed requests, defaults to 1.
+   */
+  retryCount?: number
+
+  /**
+   * Optional configuration for an OpenRPC network support.
+   */
+  rpc?: {
+
+    /**
+     * How many blocks behind the head of the blockchain to tolerate before considering the node to be unhealthy. Defaults to `2`.
+     */
+    allowedBlocksBehind?: number
+
+    /**
+     * In case this is set to `true`, the SDK will not automatically load balance and failover between the available OpenRPC nodes and will use static URLs defined bellow. Defaults to `false`.
+     */
+    ignoreLoadBalancing?: boolean
+
+    /**
+     * Wait for the load balance check to find the nearest node during the initialization. Defaults to `false`.
+     */
+    waitForFastestNode?: boolean
+
+    /**
+     * If this is set to `true`, the SDK will not automatically load balance and failover between the available OpenRPC nodes and will use the fastest URL fetched during the startup. Defaults to `false`.
+     */
+    oneTimeLoadBalancing?: boolean
+
+    /**
+     * In case this url is set, all the requests to Bitcoin will be proxied to this url without automatic load balancing and failover. For now, only first elements of the array is used as a default URL.
+     */
+    bitcoin?: {
+      url: string[]
+    }
+
+    /**
+     * In case this url is set, all the requests to Litecoin will be proxied to this url without automatic load balancing and failover. For now, only first elements of the array is used as a default URL.
+     */
+    litecoin?: {
+      url: string[]
+    }
+
+    /**
+     * In case this url is set, all the requests to Ethereum will be proxied to this url without automatic load balancing and failover. For now, only first elements of the array is used as a default URL.
+     */
+    ethereum?: {
+      url: string[]
+    }
+
+    /**
+     * In case this url is set, all the requests to Polygon will be proxied to this url without automatic load balancing and failover. For now, only first elements of the array is used as a default URL.
+     */
+    polygon?: {
+      url: string[]
+    }
+
+    /**
+     * In case this url is set, all the requests to Monero will be proxied to this url without automatic load balancing and failover. For now, only first elements of the array is used as a default URL.
+     */
+    monero?: {
+      url: string[]
+    }
+  }
 }
 
 export enum Network {