+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## [1.1.0] - 2023.04.03
+
+### Added
+- Added support for [OpenRPC network](https://openrpc.io)
+
+### Removed
+- Removed dependency on Axios - replaced with Fetch API
+- Due to above, minimum Node version is now >= 18
+
+### Changed
+- `debug` config field was renamed to `verbose`
+- renamed `TatumSdk` to `TatumSDK` as a main entry point
+
 ## [1.0.1] - 2023.03.16
 
 ### Added